requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.2
pydantic[email]==2.9.1
python-dotenv==1.0.1
email-validator==2.2.0
